requirements.txt
Django==2.1.2
django-pyodbc-azure==2.1.0.0
djangorestframework==3.8.2
pkg-resources==0.0.0
pyodbc==4.0.24
pytz==2018.5
